requirement.txt
pycryptodome
requests
certifi